[combobox-announcer] Move announcer types into src
