+++ packages/wonder-blocks-announcer/src/__tests__/util/test-utilities.ts
@@ -1,4 +1,4 @@
-import type {RegionDef, PolitenessLevel} from "../../../types/announcer.types";
+import type {RegionDef, PolitenessLevel} from "../../types/announcer.types";
 
 export function createTestRegionList(
     level: PolitenessLevel,
